第三次功課

+++ app/Http/Controllers/TeacherController.php
@@ -24,14 +24,14 @@ class TeacherController extends Controller
     public function store(Request $request)
     {
         
-        // $validator = Validator::make($request->all(), [
-        //     'name' => 'required|string|max:20',
-        // ]);
+        $validator = Validator::make($request->all(), [
+            'teacherName' => 'required|string|max:20',
+        ]);
 
-        // if ($validator->fails()) {
-        //     //$messages = $validator->errors()->getMessages();
-        //     throw new APIException('驗證錯誤' , 422);
-        // }
+        if ($validator->fails()) {
+            $messages = $validator->errors()->getMessages();
+            throw new APIException('驗證錯誤' , 422);
+        }
         $storeTeacher = [
             'teacherName' => $request->get('teacherName')
         ];
@@ -41,14 +41,14 @@ class TeacherController extends Controller
 
     public function show($id)
     {
-        // try {
-        //     $showTeacherID = $this->service->getTeacherById($id);
-        // } catch (Exception $e) {
-        //     throw new APIException('找不到對應課程', 404);
-        // }
-        // return $showTeacherID;
-        $result = Teacher::find($id);
-        return $result;
+        try {
+            $showTeacherID = Teacher::find($id);
+        } catch (Exception $e) {
+            throw new APIException('找不到對應課程', 404);
+        }
+        return $showTeacherID;
+        // $result = Teacher::find($id);
+        // return $result;
     }
 
     public function update(
@@ -64,9 +64,9 @@ class TeacherController extends Controller
 
     public function destroy($id)
     {
-        //if (! $course = Course::find($id)) {
-        //    throw new APIException('課程找不到', 404);
-        //}
+        if (! $teacher = Teacher::find($id)) {
+           throw new APIException('沒有老師', 404);
+        }
         $status = Teacher::find($id)->delete($id);
         return ['success' => $status];
     }
